requirements.txt
python-dotenv
qbittorrent-api
